+++ lib/sync_command.dart
@@ -19,6 +19,7 @@ class CommandSync<TParam, TResult> extends Command<TParam, TResult> {
     required super.includeLastResultInCommandResults,
     required super.noReturnValue,
     required super.errorFilter,
+    required super.errorFilterFn,
     required super.notifyOnlyWhenValueChanges,
     required super.name,
     required super.noParamValue,