+++ lib/undoable_command.dart
@@ -60,6 +60,7 @@ class UndoableCommand<TParam, TResult, TUndoState>
     required super.includeLastResultInCommandResults,
     required super.noReturnValue,
     required super.errorFilter,
+    required super.errorFilterFn,
     required super.notifyOnlyWhenValueChanges,
     required super.name,
     required super.noParamValue,